+++ tests/test_variants_transcripts.py
@@ -28,6 +28,7 @@ from test_needed import tests_folder, tests_config, tests_data_folder
         f"{tests_data_folder}/example.ann.vcf.gz",
         f"{tests_data_folder}/example.dbnsfp.transcripts.vcf.gz",
         f"{tests_data_folder}/example.dbnsfp.no_transcripts.vcf.gz",
+        # "/Users/lebechea/howard/data/SGT2306387.final.vcf.gz",
     ],
 )
 def test_create_transcript_view(input_vcf):
@@ -109,28 +110,30 @@ def test_create_transcript_view(input_vcf):
 
         # Check table content
         query_check = f"""
-            SELECT * FROM {transcripts_table}
-            ORDER BY "#CHROM", POS, REF, ALT, transcript
+            SELECT count(*) AS count FROM {transcripts_table}
         """
         check = variants.get_query_to_df(query=query_check)
-        assert len(check) > 0
+        assert check["count"][0] > 0
 
-        # Count number of lines in transcripts table
-        count_lines = len(check)
+        # # Check table content
+        # query_check = f"""
+        #     SELECT * FROM transcripts
+        # """
+        # check_variants = variants.get_query_to_df(query=query_check)
+        # log.debug(f"check_variants: {check_variants}")
 
         # ReCreate transcript view
         transcripts_table = variants.create_transcript_view()
 
         # Check table content
-        query_check = f"""
-            SELECT * FROM {transcripts_table}
-            ORDER BY "#CHROM", POS, REF, ALT, transcript
+        query_check2 = f"""
+            SELECT count(*) as count FROM {transcripts_table}
         """
-        check = variants.get_query_to_df(query=query_check)
-        assert len(check) > 0
+        check2 = variants.get_query_to_df(query=query_check2)
+        assert check2["count"][0] > 0
 
         # Check if number of lines is the same
-        assert len(check) == count_lines
+        assert check["count"][0] == check2["count"][0]
 
 
 @pytest.mark.parametrize(
@@ -1911,7 +1914,7 @@ def test_transcripts_create_view_export(output):
                 "transcripts_column": "PZTTranscript",
                 # "transcripts_order": ["column", "file"],
             },
-            ["NR_047519", "NR_036051", "NR_047551", "NM_001005484"],
+            ["NR_047526", "NR_036051", "NR_047551", "NM_001005484"],
         ),
         (
             {
@@ -1921,7 +1924,7 @@ def test_transcripts_create_view_export(output):
                 "transcripts_column": "PZTTranscript",
                 # "transcripts_order": ["column", "file"],
             },
-            ["NR_047519", "NR_036051", "NR_047551", "NM_001005484"],
+            ["NR_047526", "NR_036051", "NR_047551", "NM_001005484"],
         ),
         (
             {
@@ -1931,7 +1934,7 @@ def test_transcripts_create_view_export(output):
                 "transcripts_column": "PZTTranscript",
                 "transcripts_order": ["column", "file"],
             },
-            ["NR_047519", "NR_036051", "NR_047551", "NM_001005484"],
+            ["NR_047526", "NR_036051", "NR_047551", "NM_001005484"],
         ),
         (
             {
@@ -1941,7 +1944,7 @@ def test_transcripts_create_view_export(output):
                 "transcripts_column": "PZTTranscript",
                 "transcripts_order": ["file", "column"],
             },
-            ["NR_047519", "NR_036266", "NM_001346897", "NM_001005484", "NR_024540"],
+            ["NR_047526", "NR_036266", "NM_001346897", "NM_001005484", "NR_024540"],
         ),
     ],
 )